Produce files with guint32 (instead of guint16)

+++ compose-parse.py
@@ -111,7 +111,7 @@ headerfile_start = """/* GTK - The GIMP Tool Kit
  * that depend on the locale or selected input method.
  */
 
-static const guint16 gtk_compose_seqs_compact[] = {"""
+static const guint32 gtk_compose_seqs_compact[] = {"""
 
 headerfile_end = """};
 
@@ -150,7 +150,7 @@ multipleseqs_file_start = """/* GTK - The GIMP Tool Kit
 """
 
 multipleseqs_file_middle = """
-static const guint16 gtk_compose_seqs_multi[] = {"""
+static const guint32 gtk_compose_seqs_multi[] = {"""
 
 multipleseqs_file_end = """};
 """
